test(cleaning): clean-up orgs, scripts refactor (#668)
https://coveord.atlassian.net/browse/CDX-821
